computing conditions
