+++ src/main/java/com/oceanprotocol/squid/api/helper/OceanInitializationHelper.java
@@ -5,12 +5,12 @@
 
 package com.oceanprotocol.squid.api.helper;
 
+import com.oceanprotocol.common.web3.KeeperService;
 import com.oceanprotocol.keeper.contracts.*;
 import com.oceanprotocol.secretstore.core.EvmDto;
 import com.oceanprotocol.secretstore.core.SecretStoreDto;
 import com.oceanprotocol.squid.api.config.OceanConfig;
 import com.oceanprotocol.squid.external.AquariusService;
-import com.oceanprotocol.common.web3.KeeperService;
 import com.oceanprotocol.squid.manager.*;
 import org.web3j.crypto.CipherException;
