+++ src/main/java/com/oceanprotocol/squid/api/impl/AgreementsImpl.java
@@ -7,9 +7,9 @@ import com.oceanprotocol.squid.manager.OceanManager;
 import com.oceanprotocol.squid.models.Account;
 import com.oceanprotocol.squid.models.DDO;
 import com.oceanprotocol.squid.models.DID;
+import com.oceanprotocol.squid.models.service.AgreementStatus;
 import com.oceanprotocol.squid.models.service.Service;
 import com.oceanprotocol.squid.models.service.types.AccessService;
-import com.oceanprotocol.squid.models.service.AgreementStatus;
 import com.oceanprotocol.squid.models.service.types.ComputingService;
 import org.web3j.crypto.Keys;
 import org.web3j.tuples.generated.Tuple2;
