+++ src/main/java/com/oceanprotocol/squid/manager/AccountsManager.java
@@ -5,9 +5,9 @@
 
 package com.oceanprotocol.squid.manager;
 
+import com.oceanprotocol.common.web3.KeeperService;
 import com.oceanprotocol.squid.exceptions.EthereumException;
 import com.oceanprotocol.squid.external.AquariusService;
-import com.oceanprotocol.common.web3.KeeperService;
 import com.oceanprotocol.squid.models.Account;
 import com.oceanprotocol.squid.models.Balance;
 import org.apache.logging.log4j.LogManager;
