+++ src/main/java/com/oceanprotocol/squid/manager/BaseManager.java
@@ -6,15 +6,13 @@
 package com.oceanprotocol.squid.manager;
 
 import com.fasterxml.jackson.core.type.TypeReference;
+import com.oceanprotocol.common.helpers.UrlHelper;
+import com.oceanprotocol.common.web3.KeeperService;
 import com.oceanprotocol.keeper.contracts.*;
 import com.oceanprotocol.secretstore.core.EvmDto;
 import com.oceanprotocol.secretstore.core.SecretStoreDto;
-import com.oceanprotocol.squid.exceptions.DDOException;
-import com.oceanprotocol.squid.exceptions.DIDFormatException;
-import com.oceanprotocol.squid.exceptions.EncryptionException;
-import com.oceanprotocol.squid.exceptions.TokenApproveException;
+import com.oceanprotocol.squid.exceptions.*;
 import com.oceanprotocol.squid.external.AquariusService;
-import com.oceanprotocol.common.web3.KeeperService;
 import com.oceanprotocol.squid.models.Account;
 import com.oceanprotocol.squid.models.DDO;
 import com.oceanprotocol.squid.models.DID;
@@ -25,9 +23,14 @@ import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.web3j.crypto.CipherException;
-import org.web3j.crypto.Credentials;
+import org.web3j.abi.EventEncoder;
+import org.web3j.abi.FunctionReturnDecoder;
+import org.web3j.abi.datatypes.Event;
+import org.web3j.abi.datatypes.Type;
 import org.web3j.crypto.Keys;
+import org.web3j.protocol.core.DefaultBlockParameterName;
+import org.web3j.protocol.core.methods.request.EthFilter;
+import org.web3j.protocol.core.methods.response.EthLog;
 import org.web3j.protocol.core.methods.response.TransactionReceipt;
 
 import java.io.IOException;
@@ -164,6 +167,61 @@ public abstract class BaseManager {
             throw new TokenApproveException(msg, e);
         }
 
+    }
+
+
+    /**
+     * Given a DID, scans the DIDRegistry events on-chain to resolve the
+     * Metadata API url and return the DDO found
+     *
+     * @param did the did
+     * @return DDO
+     * @throws EthereumException EthereumException
+     * @throws DDOException      DDOException
+     */
+    public DDO resolveDID(DID did) throws EthereumException, DDOException {
+
+        EthFilter didFilter = new EthFilter(
+                DefaultBlockParameterName.EARLIEST,
+                DefaultBlockParameterName.LATEST,
+                didRegistry.getContractAddress()
+        );
+
+        try {
+
+            final Event event = didRegistry.DIDATTRIBUTEREGISTERED_EVENT;
+            final String eventSignature = EventEncoder.encode(event);
+            didFilter.addSingleTopic(eventSignature);
+
+            String didTopic = "0x" + did.getHash();
+            didFilter.addOptionalTopics(didTopic);
+
+            EthLog ethLog;
+
+            try {
+                ethLog = getKeeperService().getWeb3().ethGetLogs(didFilter).send();
+            } catch (IOException e) {
+                throw new EthereumException("Error searching DID " + did.toString() + " onchain: " + e.getMessage());
+            }
+
+            List<EthLog.LogResult> logs = ethLog.getLogs();
+
+            int numLogs = logs.size();
+            if (numLogs < 1)
+                throw new DDOException("No events found for " + did.toString());
+
+            EthLog.LogResult logResult = logs.get(numLogs - 1);
+            List<Type> nonIndexed = FunctionReturnDecoder.decode(((EthLog.LogObject) logResult).getData(), event.getNonIndexedParameters());
+            String ddoUrl = nonIndexed.get(0).getValue().toString();
+            String didUrl = UrlHelper.parseDDOUrl(ddoUrl, did.toString());
+
+            AquariusService ddoAquariosDto = AquariusService.getInstance(UrlHelper.getBaseUrl(didUrl));
+            return ddoAquariosDto.getDDO(didUrl);
+
+        } catch (Exception ex) {
+            log.error("Unable to retrieve DDO " + ex.getMessage());
+            throw new DDOException("Unable to retrieve DDO " + ex.getMessage());
+        }
     }
