+++ src/main/java/com/oceanprotocol/squid/manager/ConditionsManager.java
@@ -1,10 +1,12 @@
 package com.oceanprotocol.squid.manager;
 
-import com.oceanprotocol.squid.external.AquariusService;
-import com.oceanprotocol.common.web3.KeeperService;
 import com.oceanprotocol.common.helpers.EncodingHelper;
+import com.oceanprotocol.common.web3.KeeperService;
+import com.oceanprotocol.squid.external.AquariusService;
+import com.oceanprotocol.squid.models.DDO;
 import com.oceanprotocol.squid.models.DID;
 import com.oceanprotocol.squid.models.service.Agreement;
+import com.oceanprotocol.squid.models.service.Service;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.web3j.crypto.Keys;
@@ -110,15 +112,27 @@ public class ConditionsManager extends BaseManager {
     public Boolean releaseReward(String agreementId, BigInteger amount) throws Exception {
 
         Agreement agreement = new Agreement(agreementStoreManager.getAgreement(EncodingHelper.hexStringToBytes(agreementId)).send());
-        Tuple2<String, String> agreementData = escrowAccessSecretStoreTemplate.getAgreementData(EncodingHelper.hexStringToBytes(agreementId)).send();
+
+        DDO ddo = resolveDID(agreement.did);
+        Service service  = ddo.getServiceByTemplate(agreement.templateId);
+
+        Tuple2<String, String> agreementData = null;
+        TransactionReceipt txReceipt;
+
+        if (service.type.equals(Service.serviceTypes.access.name())) {
+            agreementData = escrowAccessSecretStoreTemplate.getAgreementData(EncodingHelper.hexStringToBytes(agreementId)).send();
+        } else if (service.type.equals(Service.serviceTypes.computing.name())) {
+            agreementData = escrowComputeExecutionTemplate.getAgreementData(EncodingHelper.hexStringToBytes(agreementId)).send();
+        }
 
         try {
-            TransactionReceipt txReceipt = escrowReward.fulfill(EncodingHelper.hexStringToBytes(agreementId),
+             txReceipt = escrowReward.fulfill(EncodingHelper.hexStringToBytes(agreementId),
                     amount,
                     agreementData.getValue2(),
                     agreementData.getValue1(),
                     agreement.conditions.get(1),
                     agreement.conditions.get(0)).send();
+
             return txReceipt.isStatusOK();
         } catch (TransactionException e) {
             log.error("Error releasing reward for the agreement" + agreementId + e.getMessage());
