+++ src/main/java/com/oceanprotocol/squid/models/service/Agreement.java
@@ -3,9 +3,9 @@ package com.oceanprotocol.squid.models.service;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import com.oceanprotocol.squid.exceptions.DIDFormatException;
 import com.oceanprotocol.common.helpers.EncodingHelper;
 import com.oceanprotocol.common.helpers.EthereumHelper;
+import com.oceanprotocol.squid.exceptions.DIDFormatException;
 import com.oceanprotocol.squid.models.AbstractModel;
 import com.oceanprotocol.squid.models.DID;
 import com.oceanprotocol.squid.models.FromJsonToModel;
