+++ src/main/java/com/oceanprotocol/squid/models/service/types/ComputingService.java
@@ -4,9 +4,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import com.oceanprotocol.common.helpers.EncodingHelper;
 import com.oceanprotocol.common.helpers.EthereumHelper;
-import com.oceanprotocol.squid.models.DDO;
 import com.oceanprotocol.squid.models.service.Condition;
 import com.oceanprotocol.squid.models.service.Service;
 import org.web3j.crypto.Hash;
