+++ src/main/java/com/oceanprotocol/squid/models/service/types/MetadataService.java
@@ -7,7 +7,6 @@ package com.oceanprotocol.squid.models.service.types;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import com.oceanprotocol.squid.models.asset.AssetMetadata;
 import com.oceanprotocol.squid.models.service.Service;
